requirements.txt
gql==3.4.1
graphql-query==1.2.1
pymongo==4.6.1
strawberry-graphql==0.216.0